refactor: split generator facade into focused services
Extract page publication, section redirects, MDX pass construction,
and OpenAPI refresh coordination while preserving the public facade.

Split the generator regression suite by behavior and share fixture
cleanup without emitting test utilities in production builds.
